chore: code quality improvements for SonarQube compliance
- Fix S3218: Rename shadowing methods in NifConditionExpr/NifVersionExpr
- Fix CA1305: Use Append chains instead of interpolation in ScdaDecompiler
- Fix S3973: Add curly braces to nested loops
- Fix S107: Introduce WriteFileParams and GeometryFlags records
- Fix S3260: Make private record sealed
- Add suppressions for XAML event handlers (CA1822/S2325)
- Various other code quality fixes (CS8600, CA1816, S2589)

+++ tests/Xbox360MemoryCarver.Tests/Core/Minidump/MinidumpParserTests.cs
@@ -324,7 +324,7 @@ public class MinidumpParserTests
         data[4] = 0x93;
         data[5] = 0xA7; // Version
         WriteUInt32Le(data, 8, (uint)streamCount);
-        WriteUInt32Le(data, 12, (uint)directoryOffset);
+        WriteUInt32Le(data, 12, directoryOffset);
 
         // Empty stream directory entries (type 0 = unused)
         // Just leave as zeros
